task_enroll_user_in_pending_courses method is added in philu_overrides
