+++ lms/djangoapps/philu_overrides/user_api/views.py
@@ -37,7 +37,6 @@ from util.enterprise_helpers import data_sharing_consent_requirement_at_login, i
 from util.json_request import JsonResponse
 from lms.djangoapps.onboarding.models import RegistrationType, GranteeOptIn
 
-from common.djangoapps.student.views import _enroll_user_in_pending_courses
 from social_django import utils as social_utils
 from common.djangoapps.student.helpers import AccountValidationError
 from openedx.core.djangoapps.user_authn.views.register import REGISTER_USER, record_registration_attributions
@@ -445,8 +444,6 @@ def create_account_with_params_custom(request, params, is_alquity_user):
         task_send_account_activation_email.delay(data)
     else:
         registration.activate()
-        data = {'user_id': user.id}
-        task_enroll_user_in_pending_courses.delay(data)  # Enroll student in any pending courses
 
     # Immediately after a user creates an account, we log them in. They are only
     # logged in until they close the browser. They can't log in again until they click
@@ -717,8 +714,6 @@ def create_account_with_params_custom_v2(request, params, is_alquity_user):
         task_send_account_activation_email.delay(data)
     else:
         registration.activate()
-        data = {'user_id': user.id}
-        task_enroll_user_in_pending_courses.delay(data)  # Enroll student in any pending courses
 
     # Immediately after a user creates an account, we log them in. They are only
     # logged in until they close the browser. They can't log in again until they click
@@ -767,12 +762,6 @@ def get_params_for_activation_email(request, registration, user):
     }
 
     return data
-
-
-@task()
-def task_enroll_user_in_pending_courses(data):
-    user = User.objects.get(id=data['user_id'])
-    _enroll_user_in_pending_courses(user)
 
 
 # noinspection PyMethodMayBeStatic
